Added new images to AWS bucket (wmcooper.com) for colors.
Simplified the game by removing teams because it seems that the teams thing was just a
time waster for students.
Changed background colors to white.
Moved position of the buttons to the top.
Made the cards flex based on width to maintain a nicer look of the
images even though the gaps between the images changes too.